Enum id data type modified

+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/GeneralMaster/GeneralEnumaratorGroup.cs
@@ -2,7 +2,7 @@
 {
     public partial class GeneralEnumaratorGroup
     {
-        public short GeneralEnumaratorGroupId { get; set; }
+        public int GeneralEnumaratorGroupId { get; set; }
         public string EnumGroup { get; set; }
         public Nullable<int> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
